Updated DiagonalGate define to call ported rust synthesis
