+++ crates/synthesis/src/uc_gate.rs
@@ -23,7 +23,7 @@ use nalgebra::{Matrix2, MatrixView2, Vector2};
 use numpy::{PyReadonlyArray2, ToPyArray};
 
 use crate::qsd::append;
-use crate::ucrz::diagonal_gate_circuit;
+use crate::diagonal::diagonal_gate_circuit;
 use qiskit_circuit::Qubit;
 use qiskit_circuit::circuit_data::{CircuitData, CircuitDataError};
 use qiskit_circuit::operations::Param;
