+++ qiskit/circuit/library/generalized_gates/diagonal.py
@@ -26,7 +26,6 @@ from qiskit.circuit.exceptions import CircuitError
 from qiskit.circuit.annotated_operation import AnnotatedOperation, InverseModifier
 from qiskit.utils.deprecation import deprecate_func
 
-from .ucrz import UCRZGate
 
 _EPS = 1e-10
 
@@ -102,30 +101,11 @@ class DiagonalGate(Gate):
 
         super().__init__("diagonal", num_qubits, diag)
 
-    def _define(self):
-        # Since the diagonal is a unitary, all its entries have absolute value
-        # one and the diagonal is fully specified by the phases of its entries.
+    def _define(self):       
+        from qiskit._accelerate.synthesis.diagonal import py_synth_diagonal
         diag_phases = [cmath.phase(z) for z in self.params]
-        n = len(diag_phases)
-        circuit = QuantumCircuit(self.num_qubits)
-
-        while n >= 2:
-            angles_rz = []
-            for i in range(0, n, 2):
-                diag_phases[i // 2], rz_angle = _extract_rz(diag_phases[i], diag_phases[i + 1])
-                angles_rz.append(rz_angle)
-            num_act_qubits = int(math.log2(n))
-            ctrl_qubits = list(range(self.num_qubits - num_act_qubits + 1, self.num_qubits))
-            target_qubit = self.num_qubits - num_act_qubits
-
-            ucrz = UCRZGate(angles_rz)
-            circuit.append(ucrz, qargs=[target_qubit] + ctrl_qubits)
-
-            n //= 2
-        circuit.global_phase += diag_phases[0]
-
-        self.definition = circuit
-
+        self.definition = py_synth_diagonal(diag_phases, self.num_qubits)
+        
     def validate_parameter(self, parameter):
         """Diagonal Gate parameter should accept complex
         (in addition to the Gate parameter types) and always return built-in complex."""
@@ -152,12 +132,3 @@ class DiagonalGate(Gate):
         if not np.allclose(np.abs(diag), 1, atol=_EPS):
             raise CircuitError("A diagonal element does not have absolute value one.")
 
-
-def _extract_rz(phi1, phi2):
-    """
-    Extract a Rz rotation (angle given by first output) such that exp(j*phase)*Rz(z_angle)
-    is equal to the diagonal matrix with entries exp(1j*ph1) and exp(1j*ph2).
-    """
-    phase = (phi1 + phi2) / 2.0
-    z_angle = phi2 - phi1
-    return phase, z_angle
