+++ qiskit/circuit/library/generalized_gates/uc.py
@@ -27,7 +27,7 @@ from qiskit.quantum_info.operators.predicates import is_unitary_matrix
 from qiskit.circuit.quantumcircuit import QuantumCircuit
 from qiskit.circuit.exceptions import CircuitError
 from qiskit.exceptions import QiskitError
-from qiskit._accelerate import uc_gate
+from qiskit._accelerate.synthesis import uc_gate
 
 _EPS = 1e-10  # global variable used to chop very small numbers to zero
